Fix build errors and add .gitignore
- Remove misplaced architecture test files from applet projects
- Fix CalendarEvent/ICalendarEventStorage accessibility (internal → public)
- Fix Calculator Key.OemExclaim (use Shift+D1 instead)
- Fix MidiPlayer double→float casts
- Add missing 'using System.Windows' in MainWindow.Composition.cs
- Remove duplicate project references in AccessNote.csproj
- Make all screen views public for cross-assembly usage
- Fix MainMenuModuleTests duplicate closing brace
- Update CompositionRootTests for new AppletScreens dictionary API
- Add .gitignore for bin/obj directories

All 50 tests pass, 0 warnings.

+++ src/AccessNote.Applets.MidiPlayer/MidiPlayerModule.cs
@@ -182,7 +182,7 @@ internal sealed class MidiPlayerModule
             }
 
             _sequencer.Play(_loadedMeltySynthMidi!, false);
-            _sequencer.Speed = _tempoBpm / 120.0;
+            _sequencer.Speed = (float)(_tempoBpm / 120.0);
 
             _waveOut?.Dispose();
             _waveOut = new WaveOutEvent();
@@ -230,7 +230,7 @@ internal sealed class MidiPlayerModule
 
         if (_sequencer != null)
         {
-            _sequencer.Speed = _tempoBpm / 120.0;
+            _sequencer.Speed = (float)(_tempoBpm / 120.0);
         }
 
         UpdateDisplay();
